+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 ### Added
 - Support for non-32-bpp texture images (Desktop)
 
+### Fixed
+- Flip texture image vertically on Android to match Desktop behaviour
+
 ## [1.0.0-BETA1]
 ### Added
 - `GlimpseAdapter` methods: